requirements.txt
asgiref
beautifulsoup4
blinker
certifi
cffi
charset-normalizer
click
click-didyoumean
colorama
comtypes
contourpy
crispy-bootstrap5
cycler
dj-database-url
Django
django-crispy-forms
django-email
django-environ
djangorestframework
env
Flask
fonttools
geographiclib
geopy
h3
idna
install
itsdangerous
Jinja2
joypy
kiwisolver
logging
MarkupSafe
matplotlib
MouseInfo
multipledispatch
numpy
opencv-contrib-python
p
packaging
pandas
periodictable
phonenumbers
Pillow
psycopg2
PyAudio
PyAutoGUI
pycparser
pyfiglet
pygame
PyGetWindow
PyMsgBox
pyp
pyparsing
PyPDF2
pyperclip
pypiwin32
pypng
PyRect
PyScreeze
python-dateutil
PythonTurtle
pyttsx3
pytweening
pytz
pywhatkit
pywin32
qrcode
requests
scipy
six
soupsieve
SpeechRecognition
sqlparse
timezonefinder
typing_extensions
tzdata
urllib3
Werkzeug
wifi-qrcode-generator
wikipedia
winshell